poc/api/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic==2.12.3
pydantic-settings==2.11.0
python-dotenv==1.2.1
transformers==4.49.0
pillow==12.0.0
torch==2.6.0
pdf2image==1.17.0
einops==0.8.1
timm==1.0.21
torchvision==0.21.0
